requirements.txt
fastapi==0.115.8
protobuf==5.29.3
pydantic==2.10.6
pytest==8.3.4
redis==5.2.1
Requests==2.32.3
SQLAlchemy==2.0.38
streamlit==1.41.1
uvicorn==0.34.0
google-generativeai==0.8.4
python-dotenv==1.0.1
psycopg2-binary==2.9.10
openai==1.63.2
python-telegram-bot==20.0
passlib[bcrypt]==1.7.4
PyJWT==2.3.0
python-multipart==0.0.5